chore(9.2): relocate new regression tests to match test-server glob
The npm test-server script glob (jest ./server/**/*.test.js) only
resolves to 2-level paths under sh's default globbing, so the tests I
added in server/scripts/__tests__/ and server/webhooks/__tests__/ were
invisible to the user-facing verification command. Moved them up one
level (e.g. server/scripts/sendOverdueReminders.skip-log.test.js) to
match the existing server/api-util/*.test.js convention, and fixed the
../../ imports back to ../.

No code changes — test files only, same assertions.
